Changes made to the extracts
Extract Schema:
Corrected the min and max lengths where appropriate. The field
"demoEqualizedLength" can be removed if the extract always need to
contain the raw/unequivalenced ratings data. Added more entries to the
updateInvalidReason enum. Removed the brandCode and brandName from the
mandatory list, as they are not mandatory on the buy unit. Removed
estDemos from the mandatory list as the user might not select this while
extracting the data. The maximum field for estData is corrected as this
field will hold the data for ratings/Impressions/Shr/VPH data. The
maximum value is only possible for Impressions though.

Return Schema:
Added more details to the way inboundUnits & outboundUnits will match
and how the actions will be performed on these. The duplicate
transactionId is removed.